Split submitter and calculator on their own processes, moved to logging to see what's going on docker logs
