+++ .gitignore
@@ -116,3 +116,6 @@ awaiting.txt
 done.txt
 investors.txt
 checked_comments.txt
+
+.\#*
+\#*
