--- src/models/.#investments.py
@@ -1,1 +1,0 @@
-dzervas@triton.8775:1526543149
